Made changes

+++ app/src/main/java/com/vasu/rozgaar/ui/auth/profilesetup/ProfileSetup.kt
@@ -78,9 +78,9 @@ class ProfileSetup : Fragment() {
 
     private fun postUser( name :String,pinCode: String,isOrganization: Boolean,organization:String){
         var user : User = if(isOrganization){
-            User(name,phoneNumber.toInt(),isOrganization,organization)
+            User(name,phoneNumber.toLong(),isOrganization,organization)
         }else{
-            User(name,phoneNumber.toInt(),isOrganization,null)
+            User(name,phoneNumber.toLong(),isOrganization,null)
         }
         var headers : Map<String,String> = mapOf("Authorization" to "Bearer $authorizationToken")
         viewModel.postUser(headers,user)
